+++ include/plat/pc99/plat/machine/hardware.h
@@ -11,8 +11,24 @@
 #ifndef __PLAT_MACHINE_HARDWARE_H
 #define __PLAT_MACHINE_HARDWARE_H
 
-#if defined(__KERNEL_32__)
+#include <config.h>
+#include <types.h>
+#include <plat/machine.h>
+#include <plat/machine/hardware_gen.h>
+
+#if defined(X86_32)
 #include <plat/machine/hardware_32.h>
 #endif
+
+void handleReservedIRQ(irq_t irq);
+void maskInterrupt(bool_t mask, irq_t irq);
+void ackInterrupt(irq_t irq);
+irq_t getActiveIRQ(void);
+bool_t isIRQPending(void);
+void setInterruptMode(irq_t irq, bool_t levelTrigger, bool_t polarityLow);
+void resetTimer(void);
+void platAddDevices(void);
+
+void handleSpuriousIRQ(void);
 
 #endif
